move cmd to root from x/perpetuals
