+++ protocol/x/perpetuals/client/cli/query.go
@@ -30,7 +30,6 @@ func GetQueryCmd(queryRoute string) *cobra.Command {
 	cmd.AddCommand(CmdQueryPremiumSamples())
 	cmd.AddCommand(CmdQueryPremiumVotes())
 	cmd.AddCommand(CmdQueryAllLiquidityTiers())
-	cmd.AddCommand(CmdModuleNameToAddress())
 	// this line is used by starport scaffolding # 1
 
 	return cmd
