+++ protocol/x/perpetuals/client/cli/query_perpetual.go
@@ -2,10 +2,7 @@ package cli
 
 import (
 	"context"
-	"fmt"
 	"strconv"
-
-	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/flags"
@@ -76,23 +73,6 @@ func CmdShowPerpetual() *cobra.Command {
 		},
 	}
 
-	flags.AddQueryFlagsToCmd(cmd)
-
-	return cmd
-}
-
-func CmdModuleNameToAddress() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "module-name-to-address [module-name]",
-		Short: "module name to address",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			address := authtypes.NewModuleAddress(args[0])
-			fmt.Println(address.String())
-			return nil
-		},
-	}
-
-	flags.AddPaginationFlagsToCmd(cmd, cmd.Use)
 	flags.AddQueryFlagsToCmd(cmd)
 
 	return cmd
