+++ protocol/x/perpetuals/module_test.go
@@ -211,14 +211,13 @@ func TestAppModuleBasic_GetQueryCmd(t *testing.T) {
 
 	cmd := am.GetQueryCmd()
 	require.Equal(t, "perpetuals", cmd.Use)
-	require.Equal(t, 7, len(cmd.Commands()))
+	require.Equal(t, 6, len(cmd.Commands()))
 	require.Equal(t, "get-all-liquidity-tiers", cmd.Commands()[0].Name())
 	require.Equal(t, "get-params", cmd.Commands()[1].Name())
 	require.Equal(t, "get-premium-samples", cmd.Commands()[2].Name())
 	require.Equal(t, "get-premium-votes", cmd.Commands()[3].Name())
 	require.Equal(t, "list-perpetual", cmd.Commands()[4].Name())
-	require.Equal(t, "module-name-to-address", cmd.Commands()[5].Name())
-	require.Equal(t, "show-perpetual", cmd.Commands()[6].Name())
+	require.Equal(t, "show-perpetual", cmd.Commands()[5].Name())
 }
 
 func TestAppModule_Name(t *testing.T) {
